fix(discovery): cast bigint to int in materialized view queries
PostgreSQL COUNT returns bigint which can't be serialized to JSON.
Add ::int cast to trending_score in mv_trending_novels and
mv_trending_authors queries.

+++ src/modules/discovery/discovery.service.ts
@@ -256,7 +256,7 @@ export class DiscoveryService implements OnModuleInit, OnModuleDestroy {
     const rows = await this.prisma.$queryRawUnsafe<
       Array<{ id: string; trending_score: number }>
     >(
-      `SELECT id, trending_score FROM mv_trending_novels LIMIT $1`,
+      `SELECT id, trending_score::int FROM mv_trending_novels LIMIT $1`,
       limit,
     );
 
@@ -305,7 +305,7 @@ export class DiscoveryService implements OnModuleInit, OnModuleDestroy {
     const rows = await this.prisma.$queryRawUnsafe<
       Array<{ id: string; trending_score: number }>
     >(
-      `SELECT id, trending_score FROM mv_trending_authors LIMIT $1`,
+      `SELECT id, trending_score::int FROM mv_trending_authors LIMIT $1`,
       limit,
     );
